+++ CHANGELOG.md
@@ -9,13 +9,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
+* [#603](https://gitlab.com/meltano/meltano/issues/603) `meltano select` now supports raw JSON Schema as a valid Catalog
 
 ### Changes
 
 ### Fixes
 * [#658](https://gitlab.com/meltano/meltano/issues/658) Fixed the Analyze page when no models are present
+* [#603](https://gitlab.com/meltano/meltano/issues/603) Fix an issue where `meltano select` would incorrectly report properties as excluded
+* [#603](https://gitlab.com/meltano/meltano/issues/603) Fix an issue where `meltano select` incorrectly flatten nested properties
+* [#553](https://gitlab.com/meltano/meltano/issues/553) Fix an issue where running `meltano select --list` for the first time would incorrectly report properties
 
-### Breaks
+### Break
 
 
 ## 0.25.0 - (2019-05-28)